Split folder follow by role
