--- onlinelearning/src/main/resources/templates/other/Register.html
@@ -1,157 +1,0 @@
-<!DOCTYPE html>
-<html data-bs-theme="light" xmlns:th="http://www.thymeleaf.org">
-
-<head>
-  <meta charset="utf-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no" />
-  <title>Onvid</title>
-  <link rel="stylesheet" href="assets/bootstrap/css/bootstrap.min.css" />
-  <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600&amp;display=swap" />
-  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=ABeeZee&amp;display=swap" />
-  <link rel="stylesheet" th:href="@{/assets/fonts/font-awesome.min.css}" />
-  <link rel="stylesheet" th:href="@{/assets/css/styles.min.css}" />
-
-</head>
-
-<body>
-  <nav class="navbar navbar-light navbar-expand-md" style="background: rgb(227, 175, 184)">
-    <div class="container-fluid">
-      <a class="navbar-brand" href="index.html" style="
-            background: url('assets/img/logo_transparent.png') center / cover
-              no-repeat;
-            width: 120px;
-            height: 120px;
-          "></a><button data-bs-toggle="collapse" class="navbar-toggler" data-bs-target="#navcol-1">
-        <span class="visually-hidden">Toggle navigation</span><span class="navbar-toggler-icon"></span>
-      </button>
-      <div class="collapse navbar-collapse" id="navcol-1">
-        <ul class="navbar-nav ms-auto" style="font-family: 'Abril Fatface', serif">
-          <li class="nav-item">
-            <a class="nav-link active" href="index.html">Home</a>
-          </li>
-          <li class="nav-item"><a class="nav-link" href="#">About</a></li>
-          <li class="nav-item">
-            <a class="nav-link" href="Detail.html">C<strong>art</strong></a>
-          </li>
-          <li class="nav-item"></li>
-        </ul>
-        <ul class="navbar-nav text-end ms-auto">
-          <li class="nav-item"></li>
-          <li class="nav-item"></li>
-          <li class="nav-item"></li>
-          <li class="nav-item"></li>
-        </ul>
-      </div>
-    </div>
-  </nav>
-  <div>
-    <div></div>
-    <section class="py-4 py-xl-5">
-      <div class="container">
-        <div class="row mb-5">
-          <div class="col-md-8 col-xl-6 text-center mx-auto">
-            <h2>Sign UP and Start Learning</h2>
-            <p></p>
-          </div>
-        </div>
-        <div class="row d-flex justify-content-center">
-          <div class="col-md-6 col-xl-4">
-            <div class="card mb-5">
-              <div class="card-body d-flex flex-column align-items-center">
-                <form class="text-center" method="post" th:action="@{/register}" th:object="${user}">
-                  <div class="mb-3">
-                    <input class="form-control" type="email" placeholder="Email" required th:field="*{email}" />
-                    <p th:if="${#fields.hasErrors('email')}" th:errors="*{email}"></p>
-                  </div>
-                  <div class="mb-3">
-                    <input class="form-control" type="text" placeholder="Name" th:field="*{name}" required
-                      style="margin-top: 10px" />
-                    <p th:if="${#fields.hasErrors('name')}" th:errors="*{name}"></p>
-                  </div>
-                  <div class="mb-3">
-                    <input class="form-control" type="password" name="password" th:field="*{password}" required
-                      placeholder="Password" />
-                    <p th:if="${#fields.hasErrors('password')}" th:errors="*{password}"></p>
-                  </div>
-                  <div class="mb-3">
-                    <input class="form-control" type="password" name="rePassword" th:field="*{rePassword}" required
-                      placeholder="Confirm Password" style="margin-top: 10px" />
-                    <p th:if="${#fields.hasErrors('rePassword')}" th:errors="*{rePassword}"></p>
-                  </div>
-                  <div class="mb-3">
-                    <button class="btn btn-primary d-block w-100" type="submit" style="
-                          background: #f4a0c7;
-                          color: var(--bs-btn-active-border-color);
-                        ">
-                      Sign up
-                    </button>
-                  </div>
-                  <p class="text-muted">
-                    If you have account&nbsp;<a href="Login.html">Login</a>
-                  </p>
-                </form>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>
-    <nav class="navbar navbar-light navbar-expand-md py-3">
-      <div class="container">
-        <a class="navbar-brand d-flex align-items-center" href="#"><span
-            class="bs-icon-sm bs-icon-rounded bs-icon-primary d-flex justify-content-center align-items-center me-2 bs-icon"></span></a><button
-          class="navbar-toggler" data-bs-toggle="collapse">
-          <span class="visually-hidden">Toggle navigation</span><span class="navbar-toggler-icon"></span>
-        </button>
-      </div>
-    </nav>
-    <footer style="background: #edb9c2">
-      <div class="row">
-        <div class="col-sm-6 col-md-4 footer-navigation">
-          <h3></h3>
-          <p class="links" style="
-                background: url('assets/img/logo_transparent.png') center /
-                  cover no-repeat;
-                width: 200px;
-                height: 200px;
-                margin-top: 2px;
-              "></p>
-          <p class="company-name">Company Name © 2023</p>
-        </div>
-        <div class="col-sm-6 col-md-4 footer-contacts">
-          <div>
-            <span class="fa fa-map-marker footer-contacts-icon"> </span>
-            <p>
-              <span class="new-line-span">24 Truong Dinh Street</span> Ha Noi,
-              Viet Nam
-            </p>
-          </div>
-          <div>
-            <i class="fa fa-phone footer-contacts-icon"></i>
-            <p class="footer-center-info email text-start">+89 354046455</p>
-          </div>
-          <div>
-            <i class="fa fa-envelope footer-contacts-icon"></i>
-            <p><a href="#" target="_blank">minhdung@gmail.com</a></p>
-          </div>
-        </div>
-        <div class="col-md-4 footer-about">
-          <h4>About the company</h4>
-          <p>
-            Lorem ipsum dolor sit amet, consectateur adispicing elit. Fusce
-            euismod convallis velit, eu auctor lacus vehicula sit amet.
-          </p>
-          <div class="social-links social-icons">
-            <a href="#"><i class="fa fa-facebook"></i></a><a href="#"><i class="fa fa-twitter"></i></a><a href="#"><i
-                class="fa fa-linkedin"></i></a><a href="#"><i class="fa fa-github"></i></a>
-          </div>
-        </div>
-      </div>
-      <div></div>
-    </footer>
-  </div>
-  <script src="assets/bootstrap/js/bootstrap.min.js"></script>
-  <script src="assets/js/script.min.js"></script>
-</body>
-
-</html>
