+++ onlinelearning/src/main/resources/templates/student/Register.html
@@ -1,0 +1,63 @@
+<th:block th:replace="/layout.html">
+  <th:block th:fragment="main-content">
+
+    <div>
+      <section class="py-4 py-xl-5">
+        <div class="container">
+          <div class="row mb-5">
+            <div class="col-md-8 col-xl-6 text-center mx-auto">
+              <h2>Sign Up and Start Learning</h2>
+              <p></p>
+            </div>
+          </div>
+          <div class="row d-flex justify-content-center">
+            <div class="col-md-6 col-xl-4">
+              <div class="card mb-5">
+                <div class="card-body d-flex flex-column align-items-center">
+                  <form class="text-center" method="post" th:action="@{/register}" th:object="${user}">
+                    <div class="mb-3">
+                      <input class="form-control" type="email" placeholder="Email" required th:field="*{email}" />
+                      <p th:if="${#fields.hasErrors('email')}" th:errors="*{email}"></p>
+                    </div>
+                    <div class="mb-3">
+                      <input class="form-control" type="text" placeholder="Name" th:field="*{name}" required
+                        style="margin-top: 10px" />
+                      <p th:if="${#fields.hasErrors('name')}" th:errors="*{name}"></p>
+                    </div>
+                    <div class="mb-3">
+                      <input class="form-control" type="password" name="password" th:field="*{password}" required
+                        placeholder="Password" />
+                      <p th:if="${#fields.hasErrors('password')}" th:errors="*{password}"></p>
+                    </div>
+                    <div class="mb-3">
+                      <input class="form-control" type="password" name="rePassword" th:field="*{rePassword}" required
+                        placeholder="Confirm Password" style="margin-top: 10px" />
+                      <p th:if="${#fields.hasErrors('rePassword')}" th:errors="*{rePassword}"></p>
+                    </div>
+                    <div class="mb-3">
+                      <button class="btn btn-primary d-block w-100" type="submit">
+                        Sign up
+                      </button>
+                    </div>
+                    <p class="text-muted">
+                      If you have account&nbsp;<a href="Login.html">Login</a>
+                    </p>
+                  </form>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </section>
+      <nav class="navbar navbar-light navbar-expand-md py-3">
+        <div class="container">
+          <a class="navbar-brand d-flex align-items-center" href="#"><span
+              class="bs-icon-sm bs-icon-rounded bs-icon-primary d-flex justify-content-center align-items-center me-2 bs-icon"></span></a><button
+            class="navbar-toggler" data-bs-toggle="collapse">
+            <span class="visually-hidden">Toggle navigation</span><span class="navbar-toggler-icon"></span>
+          </button>
+        </div>
+      </nav>
+    </div>
+  </th:block>
+</th:block>
